refactor: Header 컴포넌트 상태, 뷰 로직 분리
