+++ frontend/src/components/Header/hooks.js
@@ -1,0 +1,15 @@
+import { useRecoilState } from 'recoil';
+
+import { shapedDateState } from '../../recoil/date/selector';
+
+export const useShapedDate = () => {
+    const [shapedDate, setShapedDate] = useRecoilState(shapedDateState);
+
+    const changeDate = (sign) => {
+        const [NEXT, PREV] = [1, -1];
+        const value = sign === '+' ? NEXT : PREV;
+        setShapedDate(value);
+    };
+
+    return [shapedDate, changeDate];
+};
